Backend/login-service/authorization-service/requirements.txt
fastapi
python-dotenv
python-jose[cryptography]
uvicorn
PyJWT
